requirements.txt
altair==4.1.0
appnope==0.1.2
argon2-cffi==20.1.0
astor==0.8.1
async-generator==1.10
attrs==21.2.0
backcall==0.2.0
base58==2.1.0
bleach==3.3.0
blinker==1.4
cachetools==4.2.2
certifi==2021.5.30
cffi==1.14.5
chardet==4.0.0
chronometer==1.0
click==7.1.2
configparser==5.0.2
datasets==1.8.0
decorator==5.0.9
defusedxml==0.7.1
dill==0.3.3
docker-pycreds==0.4.0
entrypoints==0.3
filelock==3.0.12
fsspec==2021.6.0
future==0.18.2
gitdb==4.0.7
GitPython==3.1.17
huggingface-hub==0.0.8
idna==2.10
ipykernel==5.5.5
ipython==7.24.1
ipython-genutils==0.2.0
ipywidgets==7.6.3
jedi==0.18.0
Jinja2==3.0.1
joblib==1.0.1
jsonschema==3.2.0
jupyter-client==6.1.12
jupyter-core==4.7.1
jupyterlab-pygments==0.1.2
jupyterlab-widgets==1.0.0
MarkupSafe==2.0.1
matplotlib-inline==0.1.2
mistune==0.8.4
monotonic==0.1
multiprocess==0.70.11.1
nbclient==0.5.3
nbconvert==6.0.7
nbformat==5.1.3
nest-asyncio==1.5.1
notebook==6.4.0
numpy==1.20.3
oauthlib==3.1.1
packaging==20.9
pandas==1.2.4
pandocfilters==1.4.3
parso==0.8.2
pathtools==0.1.2
pexpect==4.8.0
pickleshare==0.7.5
Pillow==8.2.0
prometheus-client==0.11.0
promise==2.3
prompt-toolkit==3.0.18
protobuf==3.17.3
psutil==5.8.0
ptyprocess==0.7.0
pyarrow==3.0.0
pycparser==2.20
pydeck==0.6.2
Pygments==2.9.0
pymongo==3.11.4
pyparsing==2.4.7
pyrsistent==0.17.3
python-dateutil==2.8.1
python-twitter==3.5
pytz==2021.1
PyYAML==5.4.1
pyzmq==22.1.0
redis==3.5.3
regex==2021.4.4
requests==2.25.1
requests-oauthlib==1.3.0
sacremoses==0.0.45
scikit-learn==0.24.2
scipy==1.6.3
Send2Trash==1.5.0
sentencepiece==0.1.95
sentry-sdk==1.1.0
seqeval==1.2.2
Shapely==1.7.1
shortuuid==1.0.1
simpletransformers==0.61.6
six==1.16.0
smmap==4.0.0
streamlit==0.82.0
subprocess32==3.5.4
tensorboardX==2.2
terminado==0.10.1
testpath==0.5.0
threadpoolctl==2.1.0
tokenizers==0.10.3
toml==0.10.2
toolz==0.11.1
torch==1.8.1
tornado==6.1
tqdm==4.49.0
traitlets==5.0.5
transformers==4.6.1
typing-extensions==3.10.0.0
tzlocal==2.1
urllib3==1.26.5
validators==0.18.2
wandb==0.10.32
wcwidth==0.2.5
webencodings==0.5.1
widgetsnbextension==3.5.1
xxhash==2.0.2
